requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.9.0
django-heroku==0.0.0
djangorestframework==3.12.4
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
